remove stale code
