+++ scormxblock/__init__.py
@@ -1,1 +1,1 @@
-from .scromxblockng import ScormXBlock
+from .scormxblock import ScormXBlock
